statistical inference on greta models

Carry out statistical inference on greta models by MCMC or likelihood/posterior optimisation.

mcmc(model, sampler = hmc(), n_samples = 1000, thin = 1,
  warmup = 1000, chains = 1, n_cores = NULL, verbose = TRUE,
  pb_update = 50, one_by_one = FALSE, initial_values = NULL)

stashed_samples()

extra_samples(draws, n_samples = 1000, thin = 1, n_cores = NULL,
  verbose = TRUE, pb_update = 50, one_by_one = FALSE)

opt(model, optimiser = bfgs(), max_iterations = 100,
  tolerance = 1e-06, initial_values = NULL)

Arguments

model
greta_model object

sampler
sampler used to draw values in MCMC. See samplers for options.

n_samples
number of MCMC samples to draw per chain (after any warm-up, but before thinning)

thin
MCMC thinning rate; every thin samples is retained, the rest are discarded

warmup
number of samples to spend warming up the mcmc sampler. During this phase the sampler moves toward the highest density area and tunes sampler hyperparameters.

chains
number of MCMC chains to run

n_cores
the maximum number of CPU cores used by each chain.

verbose
whether to print progress information to the console

pb_update
how regularly to update the progress bar (in iterations)

one_by_one
whether to run TensorFlow MCMC code one iteration at a time, so that greta can handle numerical errors as 'bad' proposals (see below).

initial_values
an optional vector (or list of vectors, for multiple chains) of initial values for the free parameters in the model. These will be used as the starting point for sampling/optimisation.

draws
an mcmc.list object returned by mcmc or stashed_samples

optimiser
an optimiser object giving the optimisation algorithm and parameters See optimisers.

max_iterations
the maximum number of iterations before giving up

tolerance
the numerical tolerance for the solution, the optimiser stops when the (absolute) difference in the joint density between successive iterations drops below this level

Value

mcmc, stashed_samples & extra_samples - an mcmc.list object that can be analysed using functions from the coda package. This will contain mcmc samples of the greta arrays used to create model.

opt - a list containing the following named elements:

par the best set of parameters found

value the log joint density of the model at the parameters par

iterations the number of iterations taken by the optimiser

convergence an integer code, 0 indicates successful completion, 1 indicates the iteration limit max_iterations had been reached

Details

For mcmc() if verbose = TRUE, the progress bar shows the number of iterations so far and the expected time to complete the phase of model fitting (warmup or sampling). Updating the progress bar regularly slows down sampling, by as much as 9 seconds per 1000 updates. So if you want the sampler to run faster, you can change pb_update to increase the number of iterations between updates of the progress bar, or turn the progress bar off altogether by setting verbose = FALSE.

Occasionally, a proposed set of parameters can cause numerical instability (I.e. the log density or its gradient is NA, Inf or -Inf); normally because the log joint density is so low that it can't be represented as a floating point number. When this happens, the progress bar will also display the proportion of proposals so far that were 'bad' (numerically unstable) and therefore rejected. If you're getting a lot of numerical instability, you might want to manually define starting values to move the sampler into a more reasonable part of the parameter space.

Multiple mcmc chains can be run in parallel by setting the execution plan with the future package. Only plan(multisession) futures or plan(cluster) futures that don't use fork clusters are allowed, since forked processes conflict with tensorflow's parallelism.

If n_cores = NULL and mcmc chains are being run sequentially, each chain will be allowed to use all CPU cores. If chains are being run in parallel, n_cores will be set so that n_cores * chains is less than the number of CPU cores.

If the sampler is aborted before finishing, the samples collected so far can be retrieved with stashed_samples(). Only samples from the sampling phase will be returned.

Samples returned by mcmc() and stashed_samples() can be added to with extra_samples(). This continues the chain from the last value of the previous chain and uses the same sampler and model as was used to generate the previous samples. It is not possible to change the sampler or extend the warmup period.

Examples

# NOT RUN {
# define a simple model
mu <- variable()
sigma <- lognormal(1, 0.1)
x <- rnorm(10)
distribution(x) <- normal(mu, sigma)
m <- model(mu, sigma)

# carry out mcmc on the model
draws <- mcmc(m,
              n_samples = 100,
              warmup = 10)
# }# NOT RUN {
# find the MAP estimate
opt_res <- opt(m)
# }