Helpers to remove, and reinstall python environments and miniconda

This can be useful when debugging greta installation to get to "clean slate". There are five functions:

Usage

remove_greta_env(ask = interactive())

reinstall_greta_env(timeout = 5, ask = interactive())

remove_miniconda(ask = interactive())

reinstall_miniconda(timeout = 5, ask = interactive())

remove_reticulate_uv_cache(ask = interactive())

Arguments

ask

Ask for confirmation? Default is interactive().

timeout

time in minutes to wait until timeout (default is 5 minutes)

Value

Invisibly, TRUE if anything was removed, otherwise FALSE.

Details

remove_greta_env() removes the 'greta-env-tf2' conda environment

remove_miniconda() removes miniconda installation

remove_reticulate_uv_cache() removes reticulate's managed uv cache. Note this cache is shared by all R packages that use reticulate's uv (it is not greta-specific), and a system-wide uv cache is left untouched.

reinstall_greta_env() remove 'greta-env-tf2' and reinstall it using greta_create_conda_env() (which is used internally).

reinstall_miniconda() removes miniconda and reinstalls it using greta_install_miniconda() (which is used internally)

To remove everything at once, see greta_remove_all_deps().

See also

Examples

if (FALSE) { # \dontrun{
remove_greta_env()
remove_miniconda()
reinstall_greta_env()
reinstall_miniconda()
} # }