Helpers to remove, and reinstall python environments and miniconda

This can be useful when debugging greta installation to get to "clean slate". There are four functions:

remove_greta_env()

reinstall_greta_env(timeout = 5)

remove_miniconda()

reinstall_miniconda(timeout = 5)

Arguments

timeout
time in minutes to wait until timeout (default is 5 minutes)

Value

invisible

Details

remove_greta_env() removes the 'greta-env' conda environment

remove_miniconda() removes miniconda installation

reinstall_greta_env() remove 'greta-env' and reinstall it using greta_create_conda_env() (which is used internally).

reinstall_miniconda() removes miniconda and reinstalls it using greta_install_miniconda() (which is used internally)

Examples

if (FALSE) {
remove_greta_env()
remove_miniconda()
reinstall_greta_env()
reinstall_miniconda()
}